+++ core/CHANGELOG.md
@@ -1,2 +1,3 @@
 # TBD
 * Stop the node network after the test controller runs
+* Rename ServiceFactory(,Config) -> ServiceInitializer(,Core)